rag_service/requirements.txt
fastapi
uvicorn
python-multipart
requests
aiohttp
chromadb
sentence-transformers
python-dotenv
pdfplumber
pytesseract>=0.3.10
python-docx
pymupdf>=1.23.0
pdf2image>=1.16.0
cloudscraper>=1.2.71
beautifulsoup4>=4.12.0
langchain>=0.1.0
langchain-community>=0.0.20
google-generativeai
